refactoring: move get_default_params to misc.py
